Make namespace use lowercase_underscore notation
This is what well known C++ projects like BOOST, C++ Standard Library.
While of course it's mostly a matter of taste, this approach has the
benefit that one can immediately tell which parts of:

    o2:subproject::SomeClass::someMethod

are namespaces, which parts are a class and which parts are a method. Longer
discussion at:

    https://github.com/AliceO2Group/AliceO2/issues/239