Adapt Pure Interface guidelines
This reestablishes the original intention of the Google Coding Guidelines when discussing [Pure interfaces](https://google.github.io/styleguide/cppguide.html#Interfaces). I.e. the `Interface` suffix should not be considered a requirement for so called "Pure interfaces" but usage of the `Interface` suffix can be done only when certain requirements are matched.